backend/pyproject.toml
[project]
name = "kore.ai-backend"
version = "0.1.0"
description = "Enterprise Knowledge Search Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "chromadb>=1.5.5",
    "langchain>=1.2.12",
    "langchain-text-splitters>=0.3.0",
    "google-genai>=1.0.0",
    "beautifulsoup4>=4.14.3",
    "sentence-transformers>=5.2.3",
]
